@@ -4,9 +4,9 @@
   "private": true,
   "dependencies": {
     "bulma": "^0.9.1",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-router-dom": "^5.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.1",
     "sass": "^1.53.0",
     "web-vitals": "^2.1.4"
